+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* (breaking) Require gradle 5 or above (stable publishing automatic enabling removed as 5.0 enables it by default)
+
 ### 1.3.0 (2018-07-09)
 * Support new (gradle 4.8) [publishing behaviour](https://docs.gradle.org/4.8/userguide/publishing_maven.html#publishing_maven:deferred_configuration):
     - Plugin requires gradle 4.6 or above (will fail on earlier gradle).